fixes compilation error

+++ lesson-4/src/main/java/com/edi/learn/axon/query/entries/ProductEntry.java
@@ -57,7 +57,7 @@ public class ProductEntry {
         this.stock = stock;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
